refactor(ppa): move tests out into integration test dir
